% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_oauth_client.R
\name{build_oauth_client}
\alias{build_oauth_client}
\title{Build an OAuth2 client for a token endpoint}
\usage{
build_oauth_client(client_id, client_secret, token_url)
}
\arguments{
\item{client_id}{OAuth 2.0 client ID}

\item{client_secret}{OAuth 2.0 client secret}

\item{token_url}{Full URL of the OAuth 2.0 token endpoint}
}
\value{
An httr2 oauth_client object
}
\description{
Build an OAuth2 client for a token endpoint
}
\keyword{internal}
