% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphql_request.R
\name{graphql_request}
\alias{graphql_request}
\title{Execute a single GraphQL request}
\usage{
graphql_request(client, endpoint, query_string, variables)
}
\arguments{
\item{client}{An httr2 oauth_client (from \code{\link[=build_oauth_client]{build_oauth_client()}})}

\item{endpoint}{Full URL of the GraphQL endpoint}

\item{query_string}{GraphQL query as a string}

\item{variables}{Named list of GraphQL variables}
}
\value{
Parsed JSON response body as an R list
}
\description{
Execute a single GraphQL request
}
\keyword{internal}
