% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_nodes.R
\name{rename_nodes}
\alias{rename_nodes}
\title{Rename columns of a data frame according to a mapping}
\usage{
rename_nodes(df, node_names)
}
\arguments{
\item{df}{A data frame}

\item{node_names}{Named character vector where names are current column names
and values are the desired new names. Pass \code{NULL} to
return \code{df} unchanged.}
}
\value{
\code{df} with columns renamed where a mapping exists
}
\description{
Rename columns of a data frame according to a mapping
}
\keyword{internal}
