% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_graphql_inputs.R
\name{validate_graphql_inputs}
\alias{validate_graphql_inputs}
\title{Validate common GraphQL function inputs}
\usage{
validate_graphql_inputs(client_id, client_secret, time, ids, batch_size)
}
\arguments{
\item{client_id}{OAuth 2.0 client ID}

\item{client_secret}{OAuth 2.0 client secret}

\item{time}{ISO 8601 datetime string}

\item{ids}{Character vector of UUIDs}

\item{batch_size}{Positive integer}
}
\description{
Stops with an informative message when any standard argument is invalid.
Individual query functions may add their own extra checks on top.
}
\keyword{internal}
